refactor: organizando pacotes.
