+++ build.gradle
@@ -16,6 +16,7 @@ repositories {
 }
 
 dependencies {
+    implementation 'org.xerial:sqlite-jdbc:3.41.2.2'
     testImplementation 'junit:junit:4.13.2'
 }
