+++ src/main/java/com/github/jvtopsilva090/CalculadoraCPLApplication.java
@@ -1,0 +1,9 @@
+package com.github.jvtopsilva090;
+
+import com.github.jvtopsilva090.frame.PaginaPrincipal;
+
+public class CalculadoraCPLApplication {
+    public static void main(String[] args) {
+        new PaginaPrincipal().setVisible(true);
+    }
+}
