--- src/main/java/com/github/jvtopsilva090/ProjetoMatematico/principal/.gitignore
@@ -1,3 +1,0 @@
-/CalculadoraCompleta$1.class
-/CalculadoraCompleta.class
-/PaginaPrincipal.class
