+++ src/main/java/com/github/jvtopsilva090/frame/CalculadoraCompleta.java
@@ -1,4 +1,6 @@
-package principal;
+package com.github.jvtopsilva090.frame;
+
+import com.github.jvtopsilva090.service.HistoricoService;
 
 import javax.swing.*;
 import java.awt.*;
