+++ src/main/java/com/github/jvtopsilva090/service/AvaliadorExpressao.java
@@ -1,4 +1,4 @@
-package principal;
+package com.github.jvtopsilva090.service;
 
 import javax.script.ScriptEngineManager;
 import javax.script.ScriptEngine;
