+++ src/main/java/com/github/jvtopsilva090/service/HistoricoManager.java
@@ -1,4 +1,6 @@
-package principal;
+package com.github.jvtopsilva090.service;
+
+import com.github.jvtopsilva090.repository.BancoDeDados;
 
 import java.util.List;
